pyproject.toml
[tool.poetry]
name = "spc-services"
version = "0.1.0"
description = "Smart Parking Credits Services"
authors = ["Smart Parking Credits Team"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "contracts" },
    { include = "client" },
    { include = "parkingmachine" },
]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^6.15.1"
py-solc = "^3.2.0"
opencv-python = "^4.9.0.80"
opencv-python-headless = "^4.9.0.80"
typer = "^0.9.0"
python-dotenv = "^1.0.1"
json5 = "^0.9.17"


[tool.poetry.group.dev.dependencies]
web3 = "^6.15.1"
py-solc-x = "^2.0.2"
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
